Add files via upload
Student Grade Calculator

The "studentgradecalc" program is designed to calculate the total marks, average percentage, and assign a corresponding grade based on the marks obtained by a student in different subjects.

Features:

Input: The program prompts the user to enter the total number of subjects and the marks obtained in each subject (out of 100).

Calculate Total Marks: It calculates the total marks by summing up the marks obtained in all subjects.

Calculate Average Percentage: The program computes the average percentage by dividing the total marks by the total number of subjects.

Grade Assignment: Based on the average percentage achieved, the program assigns a grade to the student. The grading criteria include grades such as 'O' (Outstanding), 'A', 'B', 'C', and 'F' (Fail).

Display Results: The program displays the total marks, average percentage, and the corresponding grade to the user.

Usage:

Run the program.
Enter the total number of subjects and the marks obtained in each subject when prompted.
The program will calculate the total marks, average percentage, and assign a grade based on the input.
The results will be displayed to the user.